Location ID has been updated to reflect the following: The location ID will now be composed of: Section (2 characters 0A, 0B, ..., ZY, ZZ) Level (2 characters 0A, 0B, ..., ZY, ZZ) Vertical (2 characters 00, 01, ..., 99) Horizontal (2 characters 00, 01, ..., 99) A location ID might look like: "0A-0B-00-00" or "AZ-0Z-99-99" A location ID may either pertain to a cubby or a shelf. A cubby will require the vertical and horizontal fields to be filled in and unique. A shelf will not require the vertical and horizontal fields to be filled in. The last two new rules meant that anything pertaining to the searching or CRUD of products had to be refactored. Thorough testing needs to be done to ensure that the new in-place rules work.
Location_id is now removed from the database and replaced with individual fields of section, level, vertical, and horizontal. Prev location_id values were migrated to the new fields.

Logic must be ran on min and max quantity in the form where, if the min quantity is larger than max, return an error.

Handled DatabaseIntegrity errors with grace through middleware and a custom 500.html page

A min product Report page where gena can see products under the min zone or are in a warning threshold.